kNN algorithm from scratch!
In this file we're using an Abalone Dataset (typically a table) with information about abalones' sex, physical measurements and number of rings! With the help of the given information, using the known abalones' physical measurements and their rings, we're predicting the number of rings on abalones that are NOT in the Dataset according to their physical measurements!