Exercising kNN and SVM classifiers with a wine dataset from scikit-learn
According to the tests I made, SVM is much more for this dataset than kNN, as the Accuracy for this case is around 0.98 and the RMSE is around 0.13! As for the kNN algorithm, with the same data, with an optimized kNN we've got Accuracy of 0.71 and RMSE of 0.65! The difference is slight, but still, it's a bit better to use SVM for this dataset!