Splitted SVM on Regressor and Classifier (wine dataset)
The Regressor performed a bit worse than the Classifier, but still better than kNN! Although, in model fitting, the SVM Regressor's performance is the worst of all algorithms' I've tested for this dataset! Fitting takes around 5 seconds (+- a bit)!

I've also tried to round the results of the Regressors, but they couldn't get any close to SVC (kNN Regressor was even getting worse with rounding, lol)!

SVM Classifier is the best algorithm for this dataset! It is really fast and guesses the answers the most of all algorithms I've tested!